refactor: resource processor 위치 변경
