+++ tomcat/src/test/java/org/apache/catalina/container/ResourceProcessorTest.java
@@ -1,6 +1,13 @@
-package org.apache.coyote.http11;
+package org.apache.catalina.container;
 
-import org.apache.catalina.Manager;
+import jakarta.controller.Controller;
+import jakarta.controller.ResourceFinder;
+import jakarta.controller.StaticResourceController;
+import jakarta.http.Header;
+import jakarta.http.HttpBody;
+import jakarta.http.HttpRequest;
+import jakarta.http.HttpSessionWrapper;
+import jakarta.http.HttpVersion;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -17,12 +24,13 @@ class ResourceProcessorTest {
     @DisplayName("정적 데이터를 처리한다.")
     void processResponse() throws Exception {
         StaticResourceController resourceController = mock(StaticResourceController.class);
+        ResourceFinder resourceFinder = mock(ResourceFinder.class);
         Controller controller = mock(Controller.class);
         Controller defaultController = mock(Controller.class);
         HashMap<String, Controller> map = new HashMap<>();
         map.put("/login", controller);
         RequestMapping requestMapping = new RequestMapping(map, defaultController);
-        ResourceProcessor resourceProcessor = new ResourceProcessor(requestMapping, resourceController);
+        ResourceProcessor resourceProcessor = new ResourceProcessor(requestMapping, resourceFinder, resourceController);
 
         resourceProcessor.processResponse(createHttpRequest("GET /sample.txt HTTP/1.1"));
 
@@ -33,12 +41,13 @@ class ResourceProcessorTest {
     @DisplayName("동적 데이터를 처리한다.")
     void processDynamic() throws Exception {
         StaticResourceController resourceController = mock(StaticResourceController.class);
+        ResourceFinder resourceFinder = mock(ResourceFinder.class);
         Controller controller = mock(Controller.class);
         Controller defaultController = mock(Controller.class);
         HashMap<String, Controller> map = new HashMap<>();
         map.put("/login", controller);
         RequestMapping requestMapping = new RequestMapping(map, defaultController);
-        ResourceProcessor resourceProcessor = new ResourceProcessor(requestMapping, resourceController);
+        ResourceProcessor resourceProcessor = new ResourceProcessor(requestMapping, resourceFinder, resourceController);
 
         resourceProcessor.processResponse(createHttpRequest("GET /login HTTP/1.1"));
 
@@ -49,12 +58,13 @@ class ResourceProcessorTest {
     @DisplayName("아무런 처리도 불가능하면 기본 컨트롤러로 데이터를 처리한다.")
     void processDefault() throws Exception {
         StaticResourceController resourceController = mock(StaticResourceController.class);
+        ResourceFinder resourceFinder = mock(ResourceFinder.class);
         Controller controller = mock(Controller.class);
         Controller defaultController = mock(Controller.class);
         HashMap<String, Controller> map = new HashMap<>();
         map.put("/login", controller);
         RequestMapping requestMapping = new RequestMapping(map, defaultController);
-        ResourceProcessor resourceProcessor = new ResourceProcessor(requestMapping, resourceController);
+        ResourceProcessor resourceProcessor = new ResourceProcessor(requestMapping, resourceFinder, resourceController);
 
         resourceProcessor.processResponse(createHttpRequest("GET /unknown HTTP/1.1"));
 
@@ -62,10 +72,11 @@ class ResourceProcessorTest {
     }
 
     private HttpRequest createHttpRequest(String requestLine) {
-        return HttpRequest.createHttp11Request(requestLine,
+        return HttpRequest.createHttpRequest(requestLine,
                 Header.empty(),
                 mock(HttpBody.class),
-                mock(Manager.class)
+                HttpVersion.HTTP_1_1,
+                mock(HttpSessionWrapper.class)
         );
     }
 }
